requirements.txt
faiss==1.12.0
keyboard==0.13.5
numpy==2.4.2
opencc==1.2.0
Pillow==12.1.1
pyperclip==1.11.0
pywin32==311
sentence_transformers==5.2.3
